+++ react/form/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @einfach/react-form
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @einfach/react@0.3.2
+  - @einfach/react-utils@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes