+++ react/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @einfach/react
 
+## 0.3.2
+
+### Patch Changes
+
+- 重新实现useAtomCallback支持watchParams 而不是watch方法本身
+
 ## 0.3.1
 
 ### Patch Changes